hackathons/AWSOct2025/digital-twin/requirements.txt
# Core dependencies
streamlit>=1.28.0
boto3>=1.26.0
botocore>=1.29.0
flotorch[strands]==2.1.0b1
flotorch-sdk>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
#strands>=0.1.0

# Development dependencies (optional)
#pytest>=7.0.0
#pytest-cov>=4.0.0
#black>=23.0.0
#flake8>=6.0.0
#mypy>=1.0.0
#pre-commit>=3.0.0
